more staking

+++ packages/hardhat/contracts/IIDCounter.sol
@@ -19,5 +19,5 @@
 pragma solidity ^0.8.0;
 
 interface IIDCounter {
-  function count() external view returns (uint40);
+  function count() external view returns (uint256);
 }
